separate manifest controllers
