+++ pkg/spoke/controllers/manifestcontroller/manifestwork_controller_test.go
@@ -1,4 +1,4 @@
-package controllers
+package manifestcontroller
 
 import (
 	"fmt"
